requirements.txt
PyMySQL
pytest
pony
pytest_bdd
jsonpickle
selenium
webdriver_manager
allure-pytest